refactor: extract mod request_handler
